+++ internal/diff/plan.go
@@ -5,8 +5,8 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/MagnusOpera/pgpackage/internal/model"
-	"github.com/MagnusOpera/pgpackage/internal/projectxml"
+	"github.com/MagnusOpera/pgpac/internal/model"
+	"github.com/MagnusOpera/pgpac/internal/projectxml"
 )
 
 type Options struct {