+++ internal/diff/plan_test.go
@@ -3,8 +3,8 @@ package diff
 import (
 	"testing"
 
-	"github.com/MagnusOpera/pgpackage/internal/model"
-	"github.com/MagnusOpera/pgpackage/internal/projectxml"
+	"github.com/MagnusOpera/pgpac/internal/model"
+	"github.com/MagnusOpera/pgpac/internal/projectxml"
 )
 
 func TestBuildPlanCreateAndDrop(t *testing.T) {