+++ website/docs/whats-new.md
@@ -4,7 +4,7 @@ title: What's New
 slug: /whats-new
 ---
 
-For the complete history, see the full [CHANGELOG.md](https://github.com/MagnusOpera/pgpackage/blob/main/CHANGELOG.md) on GitHub.
+For the complete history, see the full [CHANGELOG.md](https://github.com/MagnusOpera/pgpac/blob/main/CHANGELOG.md) on GitHub.
 
 ## Unreleased